requirements.txt
flake8>=6
pytest>=7
python-pptx
